Move AWSRetry out of module_utils.ec2
